+++ src/main/java/legend/game/submap/TriangleIndicator140.java
@@ -1,13 +1,13 @@
 package legend.game.submap;
 
 public class TriangleIndicator140 {
-  public int playerX_08;
-  public int playerY_0c;
+  public float playerX_08;
+  public float playerY_0c;
   public int screenOffsetX_10;
   public int screenOffsetY_14;
   public final short[] _18 = new short[20];
-  public final short[] x_40 = new short[20];
-  public final short[] y_68 = new short[20];
+  public final float[] x_40 = new float[20];
+  public final float[] y_68 = new float[20];
   public final int[] screenOffsetX_90 = new int[20];
   public final int[] screenOffsetY_e0 = new int[20];
 }
